+++ lib/nylo.dart
@@ -18,6 +18,23 @@ class Nylo {
     router!.setNyRoutes(plugin.routes());
   }
 
+  /// Allows you to add additional Router's to your project.
+  ///
+  /// file: e.g. /lib/routes/account_router.dart
+  /// NyRouter accountRouter() => nyRoutes((router) {
+  ///    Add your routes here
+  ///    router.route("/account", (context) => AccountPage());
+  ///    router.route("/account/update", (context) => AccountUpdatePage());
+  /// });
+  ///
+  /// Usage in /lib/main.dart e.g. Nylo.addRouter(accountRouter());
+  addRouter(NyRouter router) async {
+    if (this.router == null) {
+      this.router = NyRouter();
+    }
+    this.router!.setRegisteredRoutes(router.getRegisteredRoutes());
+  }
+
   /// Run to init Nylo
   static Future<Nylo> init({required router}) async {
     await dotenv.load(fileName: ".env");